[DEVOPS-391] Refactor application for babel7
